refactor: move audit-logger.ts to audit/logger.ts (Plan E)
